ReadMe Processor for Releases (#206)
Create a python script to translate `ReadMe.md` from GitHub to PyPI. This will avoid the need to create a branch to do a release.

This script is slightly dependent on the structure of the file, so if there are substantial changes to that, this script will require updating. It also assumes that a tag `v(fairlearn.__version__)` exists in the repo.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>