Updates for GroupMetricResult and GroupMetricSet (#279)
Add (in)equality operators to `GroupMetricResult` and `GroupMetricSet`, along with basic tests. These will simplify other testing in future.

Change `GroupMetricSet` so that the `groups` have to be specified as sequential integers from zero. If this is not the case then the `compute()` method will remap the supplied groups to `[0, 1, 2, ….]` and put the stringified original values into the `group_names` property. Since the keys are now sequential integers, convert the `group_names` property itself from a dictionary into a list.

Closes #275

Signed-off-by: Richard Edgar <riedgar@microsoft.com>
Signed-off-by: Abdul Hannan Kanji <akanji@umass.edu>