Rework creation of group metric sets (#333)
Make `_create_group_metric_set()` private. Also, change the arguments and innards to require dictionaries for the predictions and sensitive features. This will require users to name each, which will nudge them towards not confusing themselves.

This is a breaking change, but the routine is unlikely to have users right now.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>