Pipeline Improvements (#373)
Working to combine some installation logicMakes a number of improvements to the pipeline YAMLs:
* Common template for installing Fairlearn
* Common test template for unit tests and notebooks

The first of these enables the previously separate test templates for 'regular' builds and the release pipeline to be consolidated. The second enables the templates for unit and notebook tests to be combined. As a result, several YAML files can be deleted.

This is part of #371 but further changes are required in order to close that issue.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>