Simplify Release Pipeline (#398)
Create a new, simplified release pipeline. This is based on the suggestion of @adrinjalali , who pointed out that uploading and then installing from Test-PyPI didn't offer much advantage over telling `pip` to install from the build wheel file. This pipeline:
1. Performs a redeployment test run
1. Creates the wheel and stores it in an Artifact
1. Installs from that wheel file, and runs tests
1. Uploads to PyPI
1. Runs tests against the version uploaded to PyPI

Whether to use Test or Prod PyPI is set by the user at queue time.

Once this has been tested with an actual production release, we will be able to remove the old release pipeline, and remove the logic surrounding the `DEV_VERSION` queue variable and `.dev[n]` version suffix in `__init__.py` (and the release scripts).

Related to #371  but will not close until the aforementioned deletions have taken place.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>