Documentation: Add notebook contribution requirements (#490)
This updates the page about contributing notebooks to include a short rubric for guiding critiques with five points:
- Deployment context
- Real harms
- Sociotechnical
- Substantiated
- For developers

The intention is that this reflects the rough consensus about the team's aspirations from the meeting on 6/18 and discussion afterward. While the "How to talk and write about fairness in AI" guide is awesome, in practice it's harder to refer to during a discussion. So this minirubric is trying to be small enough that the team might practically use it within a meeting or discussion.