Tweaks to show_versions() (#540)
Some minor changes to `show_versions()`:

- Have the python dependencies sorted alphabetically
- Add a smoke test of the `show_versions()` subroutine

The test is very much a smoke test; it verifies that the `show_versions()` subroutine does not throw any errors, and also that a few expected strings are present in the output. It does not verify that the version information itself has been fetched correctly.