Versioned documentation (#551)
Use `sphinx-multiversion` to build versioned documentation. Unfortunately, the version picker is unable to cope with files which were renamed between different documentation versions.

Splits the static landing page off from the Sphinx documentation, and copies it into place as part of the overall build pipeline. The landing page points at the documentation for `master` for now. Once we have a new release, we should consider making the landing page point at the current PyPI version (the v0.4.6 docs are too sparse for that to be a good option at the current time).

Signed-off-by: Richard Edgar <riedgar@microsoft.com>