Updates for make_derived_metric (#649)
I had overlooked some comments on the previous PR for `make_derived_metric()`. This is to address those points.

The largest change is to the data preprocessing, which has been adjusted to avoid the problem of Data Leakage between the training and test sets (credit to @adrinjalali  for the actual `Pipeline` code). This fix has been ported to the other metrics example.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>

Co-authored-by: Adrin Jalali <adrin.jalali@gmail.com>