Fix _merge_columns when using multiple sensitive features with long names (#763)
Fairlearn supports scenarios with multiple sensitive features (e.g., gender and age, but it can be more than two). Internally, the mitigation techniques currently treat this no different than a single sensitive feature, so there's a column merge step called `_merge_columns()` that happens before applying mitigations.
 
During that step, `numpy.apply_along_axis` has so far been used to merge the string representations of the sensitive features. However, this doesn't return the expected results when the names are long since it cuts them off. For example, @LeJit discovered in #728 that the results from `ThresholdOptimizer` with two sensitive features did not match the results when manually combining the two columns. Since we don't want to restrict users in sensitive feature naming this is not acceptable.

The solution in this PR is to use `pandas.DataFrame.apply` instead of `numpy.apply_along_axis` since it doesn't exhibit the same restrictions. To prevent us from hitting the same issue in the future I've also added a test case which the previous implementation would have failed.

I also suggest we release v0.6.2 once this is in since it might actually affect users (in addition to @LeJit ).

Huge kudos to @LeJit for the excellent repro in #728 . It is a rare pleasure for an engineer to get a simple repro including the specified expectations. :-)

Signed-off-by: Roman Lutz <rolutz@microsoft.com>