[DOC] Further Assessment Updates (#1129)
More tweaking of the Assessment section. The biggest change is to re-list the definitions of allocation and quality of service harms, and then expand the 'Types of Harm' portion of 'Fairness in Machine Learning' (to avoid the former simply copying the latter. And to let me reference a very good book about stereotyping harms). This continues the work identified by #1114 .

Co-authored-by: Hilde Weerts <24417440+hildeweerts@users.noreply.github.com>