Caching for MetricFrame (#1219)
Change `MetricFrame` so that all results (including aggregations) are computed in the constructor and then cached. This is a preliminary for the introduction of bootstrapping, as discussed in a number of other venues.

This may well technically be a breaking change. Although I believe that it will only change the way in which the calling code may be broken. In a nutshell, if any of the aggregation functions raise an exception, this exception has to be caught, cached and rethrown if the user invokes the aggregation which causes it. This means that if a user metric function returns an exception (for some bizarre reason), then we would raise that exception, rather than the `ValueError` which users would have received before.