MNT Fixing pandas (#1302)
## Description

The recent `pandas` 2.1 release has simultaneously introduced a new `DataFrame.map()` API, and deprecated the `DataFrame.applymap()` API which we use in `MetricFrame`. Since our builds fail on warnings, this left us with a conundrum. This PR (temporarily) solves this by changing the `DataFrame.applymap()` calls to a nested `DataFrame.apply()` pattern. This is not particularly great, but this should not be being used in performance-critical portions of `MetricFrame` and can suffice until we decide to require `pandas` be at least version 2.1.

Also make sure that Python 3.7 is fully excised and add 3.11 to the regular builds and setup.

Issue #1299 exists to track the need to tidy all of this up once we drop `pandas` v2.0 support.

## Tests

- [ ] no new tests required
- [ ] new tests added
- [x] existing tests adjusted

## Documentation

- [x] no documentation changes needed
- [ ] user guide added or updated
- [ ] API docs added or updated
- [ ] example notebook added or updated