MNT Fix Nightly Builds (#1320)
## Description

Nightly builds are failing again due to CodeCov not liking how many different results we upload. This is particularly a problem in our nightly build when we aren't getting very frequent merges to `main` (see [CodeCov notes](https://docs.codecov.com/v4.6/docs/frequently-asked-questions#is-there-a-limit-to-the-number-of-uploads-per-commit) about this issue). To ameliorate the problem, calculate the age of the current commit, and only upload the CodeCov results if the changeset is less than two days old. Each checkin to `main` should then only upload two (notionally) identical sets of results from nightly (doing it twice should protect us against Things Which Go Bump in the 'Net).

## Tests
<!--- Select all that apply by putting an x between the brackets: [x] -->
- [x] no new tests required
- [ ] new tests added
- [ ] existing tests adjusted

## Documentation
<!--- Select all that apply. -->
- [x] no documentation changes needed
- [ ] user guide added or updated
- [ ] API docs added or updated
- [ ] example notebook added or updated

## Screenshots
<!--- If your PR makes changes to visualizations (e.g., matplotlib code) or the website, please include a screenshot of the result. -->