Updating maintainers table to prepare for the v10 release.  (#1321)
## Description

The update is in setup.py but was not in generate_maintainers_table.py. I am running through the release process instructions and in checking the first step, noticed I was missing from the maintainers file. But I see I am on the v0.10.0 dev website, so let me know if this change is not needed. I am following up on the community call last week and @riedgar-ms for review :). I'll proceed with the rest of the steps when I get this sorted.

## Tests
- [X] no new tests required
- [ ] new tests added
- [ ] existing tests adjusted

## Documentation
- [X] no documentation changes needed
- [ ] user guide added or updated
- [ ] API docs added or updated
- [ ] example notebook added or updated