MNT Fix bug: FloatTransformer output dimension was set incorrectly. (#1333)
## Description
Thanks to [this elaborately typed out issue](https://github.com/fairlearn/fairlearn/issues/1332) we found a bug in Adversarial Debiasing. Considering how big, highly parametrized, and not fully covered this implementation is, I'm surprised this is the first bug :)

The bug is simple. The `FloatTransformer` class's intention is to cast different types of input to a 2D array. But, in a specific case where no transformer is provided as input, and the output is not bi/multinomial, we accidentally set the output shape as the input shape.

## Tests
<!--- Select all that apply by putting an x between the brackets: [x] -->
- [ ] no new tests required
- [ ] new tests added
- [x] existing tests adjusted

## Documentation
<!--- Select all that apply. -->
- [x] no documentation changes needed
- [ ] user guide added or updated
- [ ] API docs added or updated
- [ ] example notebook added or updated

---------

Co-authored-by: Hilde Weerts <24417440+hildeweerts@users.noreply.github.com>
Co-authored-by: Richard Edgar <riedgar@microsoft.com>